[client] | refactor(): 컴포넌트 sub 모듈 네이밍 규칙 적용
