+++ packages/client/components/CommentItem/action/make-delete-comment-action.ts
@@ -1,0 +1,7 @@
+import { Action } from 'react-fetching-library';
+
+export const makeDeleteCommentAction: Action = ({ videoId, commentId }) => ({
+  method: 'DELETE',
+  endpoint: `${process.env.API_URL_HOST}/videos/${videoId}/comments/${commentId}`,
+  credentials: 'include',
+});
