+++ packages/client/components/CommentItem/action/make-like-comment-action.ts
@@ -1,0 +1,7 @@
+import { Action } from 'react-fetching-library';
+
+export const makeLikeCommentAction: Action = ({ videoId, commentId }) => ({
+  method: 'POST',
+  endpoint: `${process.env.API_URL_HOST}/videos/${videoId}/comments/${commentId}/likes`,
+  credentials: 'include',
+});
