+++ packages/client/components/CommentItem/action/make-query-replies-action.ts
@@ -1,0 +1,7 @@
+import { Action } from 'react-fetching-library';
+
+export const makeQueryRepliesAction: Action = (videoId, commentId, page) => ({
+  method: 'GET',
+  endpoint: `${process.env.API_URL_HOST}/videos/${videoId}/comments/${commentId}/replies?page=${page}`,
+  credentials: 'include',
+});
