+++ packages/client/components/VideoComments/action/make-create-comment-action.ts
@@ -1,0 +1,8 @@
+import { Action } from 'react-fetching-library';
+
+export const makeCreateCommentAction: Action = ({ videoId, payload }) => ({
+  method: 'POST',
+  endpoint: `${process.env.API_URL_HOST}/videos/${videoId}/comments`,
+  credentials: 'include',
+  body: payload,
+});
