+++ packages/client/components/VideoComments/index.tsx
@@ -1,53 +1,63 @@
 import React from 'react';
 import InfiniteScroll from 'react-infinite-scroll-component';
 
-import * as S from './styles';
+import * as S from './style';
 import { sortOptions } from '../../constants';
 import CommentForm from '../CommentForm';
 import CommentItem from '../CommentItem';
 import { useRouter } from 'next/router';
-import { useComments } from './hooks';
 import CircularLoader from '../CircularLoader';
+import { useComments } from './hook/use-comments';
 
 const VideoComments = () => {
   const router = useRouter();
   const { videoId } = router.query;
 
-  const commentsState = useComments(videoId);
+  const {
+    count,
+    sort,
+    onSort,
+    formLoading,
+    formValue,
+    formActive,
+    onFormChange,
+    onFormFocus,
+    onFormBlur,
+    onFormCancel,
+    onFormSubmit,
+    comments,
+    onNext,
+    hasMore,
+    submittedComments,
+  } = useComments(videoId);
 
   return (
     <S.VideoComments>
-      <S.Title>
-        {commentsState.count > 0 ? `${commentsState.count}개의 댓글` : '댓글'}
-      </S.Title>
-      <S.StyledTabs
-        items={sortOptions}
-        activeValue={commentsState.sort}
-        onClick={commentsState.onSort}
-      />
+      <S.Title>{count > 0 ? `${count}개의 댓글` : '댓글'}</S.Title>
+      <S.StyledTabs items={sortOptions} activeValue={sort} onClick={onSort} />
       <CommentForm
-        loading={commentsState.formLoading}
-        value={commentsState.formValue}
-        active={commentsState.formActive}
-        onChange={commentsState.onFormChange}
-        onFocus={commentsState.onFormFocus}
-        onBlur={commentsState.onFormBlur}
-        onCancel={commentsState.onFormCancel}
-        onSubmit={commentsState.onFormSubmit}
+        loading={formLoading}
+        value={formValue}
+        active={formActive}
+        onChange={onFormChange}
+        onFocus={onFormFocus}
+        onBlur={onFormBlur}
+        onCancel={onFormCancel}
+        onSubmit={onFormSubmit}
       />
-      {commentsState.submittedComments.map(comment => (
+      {submittedComments.map(comment => (
         <CommentItem key={comment.id} {...comment} />
       ))}
       <InfiniteScroll
         style={{ overflow: 'hidden' }}
-        dataLength={commentsState.comments.length}
-        hasMore={commentsState.hasMore}
+        dataLength={comments.length}
+        hasMore={hasMore}
         loader={<CircularLoader thickness={4} />}
-        next={commentsState.onNext}
+        next={onNext}
         scrollThreshold={0.9}
       >
         <div>
-          {commentsState.comments.map(comment => (
+          {comments.map(comment => (
             <CommentItem key={comment.id} {...comment} />
           ))}
         </div>
